글 작성 기능 구현

+++ src/main/java/com/example/talent/model/user/repository/UserRepository.java
@@ -11,5 +11,5 @@ public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
     Optional<UserEntity> findByUsernameAndDeleteDateIsNull(String username);
 
-    Optional<UserEntity> findByIdAndDeleteDateIsNull(Long id);
+    Optional<UserEntity> findByIdAndDeleteDateIsNull(Integer id);
 }
